Move ExecutionPhase, ApplicationUnderstanding, ExecutionUnderstanding to bes.container package
