+++ src/edu/virginia/vcgr/genii/container/bes/execution/TerminateableExecutionPhase.java
@@ -1,7 +1,7 @@
 package edu.virginia.vcgr.genii.container.bes.execution;
 
 import edu.virginia.vcgr.genii.client.bes.ExecutionException;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 
 public interface TerminateableExecutionPhase extends ExecutionPhase
 {
