+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/CheckBinariesPhase.java
@@ -62,9 +62,8 @@ public class CheckBinariesPhase extends AbstractExecutionPhase implements Serial
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		HistoryContext history = HistoryContextFactory.createContext(HistoryEventCategory.StageIn);
 		history.createInfoWriter("Entered CheckBinariesPhase...");
 		String[] _execNameArray = _execName.split("/");
