+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/CreateWorkingDirectoryPhase.java
@@ -37,9 +37,8 @@ public class CreateWorkingDirectoryPhase extends AbstractExecutionPhase
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		HistoryContext history = HistoryContextFactory.createContext(HistoryEventCategory.CreatingActivity);
 
 		history.createTraceWriter("Creating Job Working Directory").format("Creating job working directory:  %s", _workingDirectory).close();
