+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/StageOutPhase.java
@@ -102,9 +102,8 @@ public class StageOutPhase extends AbstractExecutionPhase implements Serializabl
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		HistoryContext history = HistoryContextFactory.createContext(HistoryEventCategory.StageOut);
 
 		history.createInfoWriter("Staging %s out.", _source.getName()).format("Staging %s out to %s.", _source, _target).close();
