+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/StoreContextPhase.java
@@ -10,7 +10,7 @@ import org.morgan.util.io.StreamUtils;
 
 import edu.virginia.vcgr.genii.client.bes.ActivityState;
 import edu.virginia.vcgr.genii.client.bes.ExecutionContext;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 import edu.virginia.vcgr.genii.client.context.ICallingContext;
 import edu.virginia.vcgr.genii.container.bes.activity.BESActivity;
 
@@ -30,9 +30,8 @@ public class StoreContextPhase extends AbstractExecutionPhase implements Executi
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		
 		File callingContextFile = new File(context.getCurrentWorkingDirectory().getWorkingDirectory(), _filename);
 		FileOutputStream fos = null;
