+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudCheckStatusPhase.java
@@ -52,9 +52,8 @@ public class CloudCheckStatusPhase extends AbstractCloudExecutionPhase implement
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		CloudManager tManage = CloudMonitor.getManager(_besid);
 		String resourceID = tManage.aquireResource(_activityID);
