+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudExecutePhase.java
@@ -39,9 +39,8 @@ public class CloudExecutePhase extends AbstractCloudExecutionPhase implements Se
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		CloudManager tManage = CloudMonitor.getManager(_besid);
 		String resourceID = tManage.aquireResource(_activityID);
