+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudGenerateJobFilePhase.java
@@ -8,7 +8,7 @@ import org.ggf.bes.factory.ActivityStateEnumeration;
 
 import edu.virginia.vcgr.genii.client.bes.ActivityState;
 import edu.virginia.vcgr.genii.client.bes.ExecutionContext;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 import edu.virginia.vcgr.genii.client.jsdl.JobRequest;
 import edu.virginia.vcgr.genii.container.bes.activity.BESActivity;
 
@@ -35,9 +35,8 @@ public class CloudGenerateJobFilePhase implements ExecutionPhase, Serializable
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		FileOutputStream fOut = new FileOutputStream(_workingDir + _jobFile);
 		ObjectOutputStream oOut = new ObjectOutputStream(fOut);
 		oOut.writeObject(_job);
