+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudGenerateRunScriptPhase.java
@@ -11,7 +11,7 @@ import org.ggf.bes.factory.ActivityStateEnumeration;
 
 import edu.virginia.vcgr.genii.client.bes.ActivityState;
 import edu.virginia.vcgr.genii.client.bes.ExecutionContext;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 import edu.virginia.vcgr.genii.client.jsdl.JobRequest;
 import edu.virginia.vcgr.genii.cloud.CloudJobWrapper;
 import edu.virginia.vcgr.genii.cmdLineManipulator.config.CmdLineManipulatorConfiguration;
@@ -60,10 +60,9 @@ public class CloudGenerateRunScriptPhase implements ExecutionPhase, Serializable
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
 		try {
-			BESActivity activity = (BESActivity) activityObject;
 			FileOutputStream tStream = new FileOutputStream(_scratchDir + _runScript);
 			CloudJobWrapper.generateWrapperScript(tStream, new File(_workingDir), new File(_workingDir + _resourceFile), _job,
 				new File(_scratchDir), _manipulatorConfiguration);
