+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudProcessAccountingPhase.java
@@ -58,9 +58,8 @@ public class CloudProcessAccountingPhase extends AbstractCloudExecutionPhase imp
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 		CloudManager tManage = CloudMonitor.getManager(_besid);
 		String resourceID = tManage.aquireResource(_activityID);
