+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudSetPermissionsPhase.java
@@ -37,9 +37,8 @@ public class CloudSetPermissionsPhase extends AbstractCloudExecutionPhase implem
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 
 		CloudManager tManage = CloudMonitor.getManager(_besid);
 		String resourceID = tManage.aquireResource(_activityID);
