+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudSetupContextDirectoryPhase.java
@@ -7,7 +7,7 @@ import org.ggf.bes.factory.ActivityStateEnumeration;
 
 import edu.virginia.vcgr.genii.client.bes.ActivityState;
 import edu.virginia.vcgr.genii.client.bes.ExecutionContext;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 import edu.virginia.vcgr.genii.client.context.ClientContextResolver;
 import edu.virginia.vcgr.genii.client.context.ContextFileSystem;
 import edu.virginia.vcgr.genii.container.bes.activity.BESActivity;
@@ -31,9 +31,8 @@ public class CloudSetupContextDirectoryPhase implements ExecutionPhase, Serializ
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 
 		File dir = new File(_localWorkingDirectory);
 		dir.mkdirs();
