+++ src/edu/virginia/vcgr/genii/container/bes/execution/phases/cloud/CloudStageOutPhase.java
@@ -39,9 +39,8 @@ public class CloudStageOutPhase extends AbstractCloudExecutionPhase implements S
 	}
 
 	@Override
-	public void execute(ExecutionContext context, Object activityObject) throws Throwable
+	public void execute(ExecutionContext context, BESActivity activity) throws Throwable
 	{
-		BESActivity activity = (BESActivity) activityObject;
 
 		HistoryContext history = HistoryContextFactory.createContext(HistoryEventCategory.StageOut);
