+++ src/edu/virginia/vcgr/genii/container/jsdl/personality/common/ExecutionUnderstanding.java
@@ -1,4 +1,4 @@
-package edu.virginia.vcgr.genii.client.jsdl.personality.common;
+package edu.virginia.vcgr.genii.container.jsdl.personality.common;
 
 import java.util.Vector;
 
@@ -6,7 +6,7 @@ import org.ggf.jsdl.JobDefinition_Type;
 
 import edu.virginia.vcgr.genii.client.bes.BESConstructionParameters;
 import edu.virginia.vcgr.genii.client.jsdl.JSDLException;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 
 public interface ExecutionUnderstanding
 {
