+++ src/edu/virginia/vcgr/genii/container/jsdl/personality/forkexec/ForkExecApplicationUnderstanding.java
@@ -19,7 +19,7 @@ import edu.virginia.vcgr.genii.client.jsdl.DirBasedGridFileSystem;
 import edu.virginia.vcgr.genii.client.jsdl.JSDLException;
 import edu.virginia.vcgr.genii.client.jsdl.JSDLFileSystem;
 import edu.virginia.vcgr.genii.container.bes.activity.BESActivityServiceImpl;
-import edu.virginia.vcgr.genii.client.bes.ExecutionPhase;
+import edu.virginia.vcgr.genii.container.bes.ExecutionPhase;
 import edu.virginia.vcgr.genii.container.bes.execution.phases.PassiveStreamRedirectionDescription;
 import edu.virginia.vcgr.genii.container.bes.execution.phases.PrepareApplicationPhase;
 import edu.virginia.vcgr.genii.container.bes.execution.phases.RunProcessPhase;
